# openapi.yml
openapi: '3.1.0'
info:
  title: Comprehensive Books Management REST API
  description: >-
    **A REST API for managing Users, Books, Authors, Genres, and their interactions, developed with [Express](https://expressjs.com).**

    This API provides endpoints to manage Users, Books, Authors, Genres, and the relationships between them.
  version: 1.0.0
  contact:
    url: https://github.com/Broannn/BookNest-RestAPI
servers:
  - url: https://booknest-restapi.onrender.com/
...
paths:
...
  /books/{bookId}/genres:
    post:
      summary: Add a genre to a book.
      tags:
        - genres
      parameters:
        - name: bookId
          in: path
          required: true
          description: The ID of the book.
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                genreId:
                  type: string
                  description: The ID of the genre to add.
      responses:
        201:
          description: Genre added to the book.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookGenre'
    get:
      summary: Get genres of a book.
      tags:
        - genres
      parameters:
        - name: bookId
          in: path
          required: true
          description: The ID of the book.
          schema:
            type: string
      responses:
        200:
          description: A list of genres for the book.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Genre'
    delete:
      summary: Remove a genre from a book.
      tags:
        - genres
      parameters:
        - name: bookId
          in: path
          required: true
          description: The ID of the book.
          schema:
            type: string
        - name: genreId
          in: path
          required: true
          description: The ID of the genre to remove.
          schema:
            type: string
      responses:
        204:
          description: Genre successfully removed from the book.
...
  /genres:
    post:
      summary: Create a genre.
      tags:
        - genres
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenreData'
      responses:
        201:
          description: The genre was successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Genre'
    get:
      summary: List all genres.
      tags:
        - genres
      responses:
        200:
          description: A list of genres.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Genre'
  /genres/{genreId}:
    get:
      summary: Get a genre by ID.
      tags:
        - genres
      parameters:
        - name: genreId
          in: path
          required: true
          description: The ID of the genre.
          schema:
            type: string
      responses:
        200:
          description: The genre details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Genre'
        404:
          description: Genre not found.
    put:
      summary: Update a genre by ID.
      tags:
        - genres
      parameters:
        - name: genreId
          in: path
          required: true
          description: The ID of the genre to update.
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenreData'
      responses:
        200:
          description: Genre successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Genre'
        404:
          description: Genre not found.
    delete:
      summary: Delete a genre by ID.
      tags:
        - genres
      parameters:
        - name: genreId
          in: path
          required: true
          description: The ID of the genre to delete.
          schema:
            type: string
      responses:
        204:
          description: Genre successfully deleted.
        404:
          description: Genre not found.
...
components:
...
  schemas:
    Genre: # Fixed indentation for Genre schema
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the genre.
          example: 5fbd827a512e73129cfedb1c
        name:
          type: string
          description: The name of the genre.
          example: Fiction
        createdAt:
          type: string
          format: date-time
          description: The date the genre was created.
          example: 2025-01-13T10:00:00.000Z
      required:
        - id
        - name
    GenreData:
      type: object
      properties:
        name:
          type: string
          description: The name of the genre.
          example: Fiction
      required:
        - name
...